+++ packages/zowe-explorer-api/src/profiles/ZoweExplorerZosmfApi.ts
@@ -122,26 +122,6 @@ export class ZosmfUssApi extends ZosmfApiCommon implements ZoweExplorerApi.IUss 
         return zowe.Utilities.putUSSPayload(this.getSession(), outputPath, { ...(options ?? {}), request: "copy" });
     }
 
-    /**
-     * API method to wrap to the newer `putContent`.
-     * @deprecated
-     */
-    public putContents(
-        inputFilePath: string,
-        ussFilePath: string,
-        binary?: boolean,
-        localEncoding?: string,
-        etag?: string,
-        returnEtag?: boolean
-    ): Promise<zowe.IZosFilesResponse> {
-        return this.putContent(inputFilePath, ussFilePath, {
-            binary,
-            localEncoding,
-            etag,
-            returnEtag,
-        });
-    }
-
     public putContent(inputFilePath: string, ussFilePath: string, options: zowe.IUploadOptions): Promise<zowe.IZosFilesResponse> {
         return zowe.Upload.fileToUssFile(this.getSession(), inputFilePath, ussFilePath, options);
     }
@@ -266,10 +246,6 @@ export class ZosmfJesApi extends ZosmfApiCommon implements ZoweExplorerApi.IJes 
         return zowe.GetJobs.getJobsByParameters(this.getSession(), params);
     }
 
-    public getJobsByOwnerAndPrefix(owner: string, prefix: string): Promise<zowe.IJob[]> {
-        return zowe.GetJobs.getJobsByOwnerAndPrefix(this.getSession(), owner, prefix);
-    }
-
     public getJob(jobid: string): Promise<zowe.IJob> {
         return zowe.GetJobs.getJob(this.getSession(), jobid);
     }
@@ -312,10 +288,6 @@ export class ZosmfJesApi extends ZosmfApiCommon implements ZoweExplorerApi.IJes 
  */
 // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 export class ZosmfCommandApi extends ZosmfApiCommon implements ZoweExplorerApi.ICommand {
-    public issueTsoCommand(command: string, acctNum: string): Promise<zowe.IIssueResponse> {
-        return zowe.IssueTso.issueTsoCommand(this.getSession(), acctNum, command);
-    }
-
     public issueTsoCommandWithParms(command: string, parms: zowe.IStartTsoParms): Promise<zowe.IIssueResponse> {
         return zowe.IssueTso.issueTsoCommand(this.getSession(), parms.account, command, parms);
     }