+++ packages/zowe-explorer-api/src/security/KeytarApi.ts
@@ -22,8 +22,8 @@ export class KeytarApi {
     public async activateKeytar(initialized: boolean, isTheia: boolean): Promise<void> {
         const log = imperative.Logger.getAppLogger();
         const profiles = new ProfilesCache(log, vscode.workspace.workspaceFolders?.[0]?.uri.fsPath);
-        const scsActive = profiles.isSecureCredentialPluginActive();
-        if (scsActive) {
+        const isSecure = await profiles.isCredentialsSecured();
+        if (isSecure) {
             const keytar = KeytarCredentialManager.getSecurityModules("keytar", isTheia);
             if (!initialized && keytar) {
                 // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment