+++ packages/zowe-explorer-api/src/profiles/ZoweExplorerApi.ts
@@ -107,6 +107,14 @@ export namespace ZoweExplorerApi {
          * @returns {Promise<boolean>}
          */
         isFileTagBinOrAscii(ussFilePath: string): Promise<boolean>;
+        /**
+         * Copy operation for USS files or directories.
+         *
+         * @param {string} outputPath the output/destination path for the file/directory
+         * @param {object} options Other options for the API endpoint
+         * @returns {Promise<Buffer>}
+         */
+        copy?(outputPath: string, options?: Omit<object, "request">): Promise<Buffer>;
 
         /**
          * Retrieve the contents of a USS file.
@@ -308,6 +316,17 @@ export namespace ZoweExplorerApi {
          * @returns {Promise<zowe.IZosFilesResponse>}
          */
         dataSetsMatchingPattern?(filter: string[], options?: zowe.IDsmListOptions): Promise<zowe.IZosFilesResponse>;
+
+        /**
+         * Copies a dataSet.
+         *
+         * @param {string} fromDataSetName
+         * @param {string} toDataSetName
+         * @param {string?} enq possible values : {SHR, SHRW, EXCLU}
+         * @param {boolean?} replace
+         * @returns {Promise<zowe.IZosFilesResponse>}
+         */
+        copyDataSet?(fromDataSetName: string, toDataSetName: string, enq?: string, replace?: boolean): Promise<zowe.IZosFilesResponse>;
     }
 
     /**