+++ packages/zowe-explorer/src/command/TsoCommandHandler.ts
@@ -13,7 +13,6 @@ import * as vscode from "vscode";
 import * as nls from "vscode-nls";
 import * as globals from "../globals";
 import { Gui, ValidProfileEnum, IZoweTreeNode } from "@zowe/zowe-explorer-api";
-import { PersistentFilters } from "../PersistentFilters";
 import { Profiles } from "../Profiles";
 import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
 import { errorHandling, FilterDescriptor, FilterItem } from "../utils/ProfilesUtils";